requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
crispy-bootstrap5==2025.6
cryptography==45.0.7
cssselect2==0.8.0
daphne==4.2.1
Django==5.2.4
django-allauth==65.11.2
django-countries==7.6.1
django-crispy-forms==2.4
django-filter==25.1
django-otp==1.6.2
django-redis==6.0.0
django-tenants==3.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.60.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
msgpack==1.1.1
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyotp==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
pytz==2025.2
qrcode==8.2
redis==6.3.0
reportlab==4.4.3
requests==2.32.4
sentry-sdk==2.42.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
structlog==25.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
xlsxwriter==3.2.5
xlwings==0.33.15
zope.interface==8.0
zopfli==0.2.3.post1
